can make uibutton glow multiply (/destadd?) blend mode looked briefly at x64 build

+++ GameCommon/Console/Console.cpp
@@ -215,7 +215,7 @@ BOOL	RegisteredConsoleCommandList::Register( const char* szCommandName, ConsoleC
 {
 RegisteredConsoleCommandList*		pNewCommand = new RegisteredConsoleCommandList;
 
-	pNewCommand->mszCommandName = (char*)( SystemMalloc( strlen( szCommandName ) + 1 ) );
+	pNewCommand->mszCommandName = (char*)( SystemMalloc( (int)strlen( szCommandName ) + 1 ) );
 	strcpy( pNewCommand->mszCommandName, szCommandName );
 	pNewCommand->mfnCommandFunction = fnFunction;
